Provide a module with validation exceptions
This introduces a new module for switching the validation scheme. Instead of
using the is_valid() function that returns a boolean a validate() function
either returns the sanitised number or raises an exception that should
indicate the kind of validation failure.

This should make it easier for applications calling this library to present
more informative messages to the user.