Implement alternate VIES check
Add a function to stdnum.eu.vat so that when one does a VIES VAT number
check, one gets a proof (certificate) that one did the check, as defence
against the VAT administration later putting this in doubt. This
certificate is provided by the VIES service, if one provides one's own
VAT number.