Add Mexican RFC number
This adds support for the Mexican tax number RFC (Registro Federal de
Contribuyentes).

This module includes a number of checks on the number but the validation
of the last check digit is disabled by default because a large number of
numbers were found that were otherwise valid but had an invalid check
digit.