Fix bug in Irish VAT number validation
The last digits of the number that should be letters were not tested to
be letters which could result in ValueError being raised for certain
validations.

This also clarifies the documentation and adds a convert() function to
convert numbers from the old format (where the second character would be
a letter or symbol) to the new format (7 digits followed by 1 or 2
letters).