Add test for Ñ in Referencia Catastral
This supports the Referencia Catastral with an Ñ in it for both byte
strings (Python 2) and unicode strings (Python 2 and 3). Support for
literal unicode strings in Python 2 doctests is flaky so the test is a
bit ugly.

This also adds a few numbers that were found online. Sadly no real
numbers with an Ñ in it have been found so the one in the test was
constructed.