Ensure 100% branch coverage
This ensures that the tests fail if 100% branch coverage is not
achieved. It also adds some pragma statements for code that cannot be
covered or is Python version dependent.

Furthermore, the get_module_list() function was removed from stdnum.util
and more tests were made from stdnum.util and stdnum.numdb. The
functionality to call format() in a country-specific IBAN implementation
was also dropped because it was not used.