Correctly split IMSI with multi-length MNC
Ensures that imsi.split() will correctly split the IMSI into an MCC, MNC
and MSIN even if not all MNCs within a single MCC have the same length.
This has the downside of not being able to guess the MNC length in some
cases.

This also omits empty information from the data file and updates the
data file from Wikipedia.

Closes https://github.com/arthurdejong/python-stdnum/issues/68