Use an internal isdigits() function instead of str.isdigit()
The problem with the latter is that it will also accept all kinds of
unicode digits that are not the ASCII 0-9 digits causing all kinds of
problems in check digit calculations.

Some of these unicode characters are also considered digits by int() but
some are not (such as the SUPERSCRIPT TWO unicode character).

Closes https://github.com/arthurdejong/python-stdnum/issues/96