Support new btw-identificatienummer
The btw-identificatienummer has been introduced on January 1st 2020 in
the Netherlands as an alternative to the btw-nummer that contains the
BSN personal identifier. The number has the same structure and function
but does not contain a BSN and uses a different check digit algorithm.

Thanks to Cas Vissers, Jeroen van Heiningen, Jerome Hanke, Nicolas
Martinelli, Ronald Portier and Tim Muller for contributing to the fix.

More information:

* http://kleineondernemer.nl/index.php/nieuw-btw-identificatienummer-vanaf-1-januari-2020-voor-eenmanszaken
* https://nl.wikipedia.org/wiki/Btw-nummer_(Nederland)
* https://www.belastingdienst.nl/wps/wcm/connect/bldcontenten/belastingdienst/business/vat/new-vat-id/
* https://www.belastingdienst.nl/wps/wcm/connect/bldcontentnl/belastingdienst/zakelijk/btw/administratie_bijhouden/btw_nummers_controleren/uw_btw_nummer

Closes https://github.com/arthurdejong/python-stdnum/issues/182
Closes https://github.com/arthurdejong/python-stdnum/pull/183
Closes https://github.com/arthurdejong/python-stdnum/pull/184
Closes https://github.com/arthurdejong/python-stdnum/pull/185