Fix misleading docstring in SE get_birth_date.
The docstring for get_birth_date() in the Swedish personnummer.py warned
that the datetime might be 100 years off because of the lack of
precision in the personnummer.

This was accurate when the docstring was written, but this is no longer
accurate after @mxamin's commit 5441ffada1a8ff6e860f5675264d40bfbb470820

However, it seems like @mxamin has forgotten to update the docstring, so
this commit updates the docstring to correspond with what is actually
happening in this function after this change was made.