Fix misleading docstring in se.personnummer get_birth_date()
The docstring for get_birth_date() in the Swedish personnummer.py warned
that the datetime might be 100 years off because of the lack of
precision in the personnummer.

This was accurate when the docstring was written, but this is no longer
accurate after the - and + signs were correctly handled.

Fixes 5441ffa