Implement get_birth_date for no.fodselsnummer.
This commit adds a function that allows you to determine a persons birth
date from a Norwegian fødselsnummer.

This also accounts for D-numbers, H-numbers, and FH-numbers, which
contain special exceptions and modifications to the birthdate portion of
the number.

Most of the information this is based on was found here:
https://no.wikipedia.org/wiki/F%C3%B8dselsnummer#H-nummer