Implement get_birth_date() for no.fodselsnummer
This adds a function that allows you to determine a persons birth date
from a Norwegian fødselsnummer.

This also accounts for D-numbers, H-numbers, and FH-numbers, which
contain special exceptions and modifications to the birthdate portion of
the number.

Most of the information this is based on was found here:
https://no.wikipedia.org/wiki/F%C3%B8dselsnummer#H-nummer

It also updates the list of valid fødselsnummer in the tests, since this
list contained many numbers that are not valid by this new validation
that now accounts for dates.

Additionally, this updates all tests that were failing under the new
validation, and adds a few new tests to bring the coverage to 100%.

Closes https://github.com/arthurdejong/python-stdnum/pull/187