Update tests for no.fodselsnummer.
This documents the new validation and the new get_birth_date feature
for Norwegian fødselsnummer.

It also updates the list of valid fødselsnummer, since this list
contained many numbers that are not valid by this new validation that
now accounts for dates.

Additionally, this updates all tests that were failing under the new
validation, and adds a few new tests to bring the coverage to 100%.