Stop non-operational MNCs from confusing IMSI dataset
This only includes data from non-operational (status "Not operational"
according to Wikipedia) Mobile Network Code operators in the generated
data file if they would not confuse the lookup of operational numbers.

This avoid problems when the "030" to "039" non-operational ranges
conflicting with the "03" operational range. This ensures that only the
"03" value is kept. For historical completeness we keep the other
non-operational values.

Closes https://github.com/arthurdejong/python-stdnum/issues/257