Only process the shortest matches in the numdb module
This ensures that matching numbers is done consistently when the numdb
database file has conflicting information about the length of numbers.

This also refactors the _find() function to be simpler and reduces the
number of recursive calls that have to be done.

The tests have been re-formatted to use pprint to make it easier to spot
differences if any of the tests fail (instead of just saying expected
True, got False).

Closes https://github.com/arthurdejong/python-stdnum/issues/257