Add North Macedonian ЕДБ
Note that this is implementation is mostly based on unofficial sources
describing the format, which match the hundreds of examples found
online.
https://forum.it.mk/threads/modularna-kontrola-na-embg-edb-dbs-itn.15663/?__cf_chl_tk=Op2PaEIauip6Z.ZjvhP897O8gRVAwe5CDAVTpjx1sEo-1663498930-0-gaNycGzNCRE#post-187048

Also note that the algorithm for the check digit was tested on all found
examples, and it doesn't work for all of them, despite those failing
examples don't seem to be valid according to the official online search.

Closes https://github.com/arthurdejong/python-stdnum/pull/330
Closes https://github.com/arthurdejong/python-stdnum/issues/222