Add support for North Macedonia TIN
Note that this is implementation is mostly based on unofficial sources
describing the format, which match the hundreds of examples found online.

Also note that the algorithm for the check digit was tested on all found
examples, and it doesn't work for all of them, despite those failing examples
don't seem to be valid according to the official online search.

Closes #222.