Fix Albanian tax number validation
This extends the description of the Albanian NIPT (NUIS) number with
information on the structure of the number. The first character was
previously limited between J and L but this letter indicates a decade
and the number is also used for individuals to where it indicates a
birth date.

Thanks Julien Launois for pointing this out.

Source: https://www.oecd.org/tax/automatic-exchange/crs-implementation-and-assistance/tax-identification-numbers/Albania-TIN.pdf

Fixes 3db826c
Closes https://github.com/arthurdejong/python-stdnum/pull/402