Remove EU NACE update script
The website that publishes the NACE catalogue has changed and a complete
re-write of the script would be necessary. The data file hasn't changed
since 2017 so is also unlikely to change until it is going to be
replaced by NACE rev. 2.1 in 2025.

See https://ec.europa.eu/eurostat/web/nace

The NACE rev 2 specification can now be found here:
https://showvoc.op.europa.eu/#/datasets/ESTAT_Statistical_Classification_of_Economic_Activities_in_the_European_Community_Rev._2/data

The NACE rev 2.1 specification can now be found here:
https://showvoc.op.europa.eu/#/datasets/ESTAT_Statistical_Classification_of_Economic_Activities_in_the_European_Community_Rev._2.1._%28NACE_2.1%29/data

In both cases a ZIP file with RDF metadata can be downloaded (but the
web applciation also exposes some simpler JSON APIs).