Update database files
The Belgian bpost bank no longer has a registration and a few bank
account numbers in the tests that used that bank were removed.

Also updates the update/gs1_ai.py script to handle the new format of the
data published by GS1. Also update the GS1-128 module to handle some
different date formats.

The Pakistan entry was kept in the stdnum/iban.dat file because the PDF
version of the IBAN Registry still contains the country.

fix db