Customise certificate validation for web services
This adds a `verify` argument to all functions that use network services
for lookups. The option is used to configure how certificate validation
works, the same as in the requests library.

For SOAP requests this is implemented properly when using the Zeep
library. The implementations using Suds and PySimpleSOAP have been
updated on a best-effort basis but their use has been deprecated because
they do not seem to work in practice in a lot of cases already.

Related to https://github.com/arthurdejong/python-stdnum/issues/452
Related to https://github.com/arthurdejong/python-stdnum/pull/453