Ignore test failures from www.dgii.gov.do
There was a change in the SOAP service and there is a new URL. However,
the API has changed and seems to require authentication.

We ignore test failures for now but unless a solution is found the DGII
validation will be removed.

See: https://github.com/arthurdejong/python-stdnum/pull/462
See: https://github.com/arthurdejong/python-stdnum/issues/461